# swagger.yaml
openapi: 3.0.3
info:
  title: ServiceRoute Integration Module API
  description: API для системы управления заявками на обслуживание с интеграцией 1С
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшн сервер
...
paths:
...
  /requests:
    get:
      tags:
        - Заявки
      summary: Список всех заявок
      description: Возвращает HTML страницу со списком всех заявок
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HTML страница со списком заявок
          content:
            text/html:
              schema:
                type: string
  /requests/new:
    get:
      tags:
        - Заявки
      summary: Форма создания новой заявки
      description: Возвращает HTML форму для создания новой заявки
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HTML форма создания заявки
          content:
            text/html:
              schema:
                type: string
...
  /requests/{id}:
    get:
      tags:
        - Заявки
      summary: Просмотр заявки
      description: Возвращает HTML страницу с детальной информацией о заявке
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML страница с информацией о заявке
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Заявка не найдена
  /requests/{id}/edit:
    get:
      tags:
        - Заявки
      summary: Форма редактирования заявки
      description: Возвращает HTML форму для редактирования заявки (только для администраторов)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML форма редактирования заявки
          content:
            text/html:
              schema:
                type: string
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
        '404':
          description: Заявка не найдена
...
  /requests/{id}/reserve-parts:
    get:
      tags:
        - Заявки
      summary: Форма резервирования запчастей
      description: Возвращает HTML форму для резервирования запчастей для заявки
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML форма резервирования запчастей
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Заявка должна быть в статусе 'Назначена' или 'В работе'
        '403':
          description: Доступ запрещен (требуется роль ENGINEER)
        '404':
          description: Заявка не найдена

    post:
      tags:
        - Заявки
      summary: Резервирование запчастей для заявки
      description: Резервирует запчасти для заявки
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                partIds:
                  type: array
                  items:
                    type: integer
                  description: Список ID запчастей для резервирования
                  example: [1, 2, 3]
                quantities:
                  type: array
                  items:
                    type: integer
                  description: Количество каждой запчасти
                  example: [2, 1, 5]
      responses:
        '302':
          description: Редирект на страницу заявки
          headers:
            Location:
              schema:
                type: string
                example: "/requests/1"
        '400':
          description: Ошибка валидации данных
        '403':
          description: Доступ запрещен (требуется роль ENGINEER)
        '500':
          description: Внутренняя ошибка сервера
  /requests/{id}/complete-form:
    get:
      tags:
        - Заявки
      summary: Форма завершения заявки
      description: Возвращает HTML форму для завершения заявки и создания заказа
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML форма завершения заявки
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Заявка должна быть в статусе 'В работе'
        '403':
          description: Доступ запрещен (требуется роль ENGINEER)
        '404':
          description: Заявка не найдена
...
  /requests/{requestId}/order/new:
    get:
      tags:
        - Заказы
      summary: Форма создания заказа из завершенной заявки
      description: Возвращает HTML форму для создания заказа из завершенной заявки
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: HTML форма создания заказа
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Заявка должна быть завершена
        '404':
          description: Заявка не найдена
...
  /customers:
    get:
      tags:
        - Клиенты
      summary: Список всех клиентов
      description: Возвращает HTML страницу со списком всех клиентов
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HTML страница со списком клиентов
          content:
            text/html:
              schema:
                type: string
  /customers/create:
    get:
      tags:
        - Клиенты
      summary: Редирект на список клиентов
      description: Перенаправляет на страницу списка клиентов
      security:
        - bearerAuth: []
      responses:
        '302':
          description: Редирект на /customers
          headers:
            Location:
              schema:
                type: string
                example: "/customers"

    post:
      tags:
        - Клиенты
      summary: Создание нового клиента
      description: Создает нового клиента в системе
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - fullName
                - phoneNumber
              properties:
                fullName:
                  type: string
                  description: Полное имя клиента
                  example: "Иванов Иван Иванович"
                phoneNumber:
                  type: string
                  description: Номер телефона клиента
                  example: "+79991234567"
                redirectTo:
                  type: string
                  description: URL для редиректа после создания (опционально)
                  example: "/requests/new"
      responses:
        '302':
          description: Редирект на страницу списка клиентов или указанный URL
          headers:
            Location:
              schema:
                type: string
                example: "/customers"
        '400':
          description: Ошибка валидации данных или клиент с таким телефоном уже существует
        '500':
          description: Внутренняя ошибка сервера
  /nomenclature:
    get:
      tags:
        - Номенклатура
      summary: Список номенклатуры
      description: Возвращает HTML страницу со списком услуг и запчастей
      security:
        - bearerAuth: []
      responses:
        '200':
          description: HTML страница со списком номенклатуры
          content:
            text/html:
              schema:
                type: string
  /nomenclature/sync:
    post:
      tags:
        - Номенклатура
      summary: Синхронизация номенклатуры из 1С
      description: Синхронизирует услуги и запчасти из системы 1С
      security:
        - bearerAuth: []
      responses:
        '302':
          description: Редирект на страницу номенклатуры
          headers:
            Location:
              schema:
                type: string
                example: "/nomenclature"
        '500':
          description: Внутренняя ошибка сервера
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
...
security:
  - bearerAuth: []